20160603 (0.6.00)

+++ libmodplay/modplay.c
@@ -282,7 +282,7 @@ void MODSetMono(int value)
 
 // Some code might use this, so use a stub.
 
-int SsFreqToPitch(int hz)
+unsigned short SsFreqToPitch(int hz)
 {
 	return 0;
 }
